fle-backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
httpx
groq
fitz  # alias de PyMuPDF, installé avec pymupdf
pymupdf
sentence-transformers
faiss-cpu
youtube-search-python
streamlit

